BUG GH44079 fix

+++ pandas/io/parsers/base_parser.py
@@ -698,7 +698,7 @@ class ParserBase:
             # error: Argument 2 to "isin" has incompatible type "List[Any]"; expected
             # "Union[Union[ExtensionArray, ndarray], Index, Series]"
             mask = algorithms.isin(values, list(na_values))  # type: ignore[arg-type]
-            na_count = mask.sum()
+            na_count = mask.astype('uint8').sum()
             if na_count > 0:
                 if is_integer_dtype(values):
                     values = values.astype(np.float64)
